+++ ghost-serialization/src/commonMain/kotlin/com/ghost/serialization/parser/GhostJsonStringReaderSubsystem.kt
@@ -333,7 +333,7 @@ private fun GhostJsonStringReader.internalSelect(options: JsonReaderOptions, con
     val length = end - start
     val key = computeKeyHash(start, length, options.hasCollisions)
 
-    val hasIndex = ((key * options.multiplier + length) shr options.shift) and C.HASH_MASK
+    val hasIndex = ((key * options.multiplier + length) shr options.shift) and (options.dispatch.size - 1)
     val index = options.dispatch[hasIndex]
 
     if (index != C.MATCH_END) {